requirements.txt
streamlit
instaloader
openai-whisper
ffmpeg-python
python-dotenv
groq
requests